LODraw: default template, no more blue objects!
I can't be bothered to do this right now, but this is set as the
default template in

	${XDG_CONFIG_HOME}/libreoffice/4/user/registrymodifications.xcu

This is in the XML element

	<item oor:path="/org.openoffice.Setup/Office/Factories/org.openoffice.Setup:Factory['com.sun.star.drawing.DrawingDocument']">
		<prop oor:name="ooSetupFactoryTemplateFile" oor:op="fuse">
			<value>$(user)/template/Defaults/Draw.otg</value>
		</prop>
	</item>

I'm fairly sure dropping the template in and settings this up should
work, but I'm not certain the template manager will like it. This may
also need configuring.

In the end, if extensive modifications are made it's going to be
easier to just backup the whole user folder and drop it all in but for
now ~8kB is better than 1.8MB for a single customisation.