feat(nvim): add (global) hint filter for pyright
When using pyright & ruff-lsp, both warn about similar things resulting
in multiple similar hints. This is an experimental approach to filtering
out specific hints.

This implementation is applied to all LSPs, with the filter checking the
source is pyright before applying the filter. There is an enhanced
approach that applies the filter to a specific lsp but just banking this
as it works without refinement.