fix: some non-transparency in gruvbox
Not fixed everywhere, but the gutter is no longer grey and that by
itself makes a big difference

I am probably missing something but passing opts in at the same time as
config in the lazy spec is causing the scheme to not pick up the
transparency option. Regardless, setting the colour scheme separately
rather than against an input is more flexible if switching to others (I
prefer to have colour schemes I actually have tested and like in the
bank, inactive).